Update test results for cpp
