--- test/data/car.cpp.cpp
@@ -1,124 +1,0 @@
-/**
- * \file Vehicle.cpp
- */
-
-class Vehicle {
-
-  public:
-    Vehicle() {
-      // @todo
-    }
-
-}
-
-/**
- * \file Car.cpp
- */
-
-class Car: public Vehicle {
-  private:
-    String model;
-    String make;
-    Number year;
-
-  public:
-    Car() {
-      // @todo
-    }
-
-}
-
-/**
- * \file NamesInThings.cpp
- */
-
-class NamesInThings {
-  private:
-    String _some_private;
-    String_2 field_2;
-
-  public:
-    NamesInThings() {
-      // @todo
-    }
-  private:
-    /**
-     */
-    void member() {
-
-    }
-
-    /**
-     *  @return String1
-     */
-    String1 member2() {
-      return null;
-    }
-
-    /**
-     */
-    void member3() {
-
-    }
-
-    /**
-     *  @return String2
-     */
-    String2 member_s() {
-      return null;
-    }
-
-}
-
-/**
- * \file Toyota.cpp
- */
-
-class Toyota: public Car {
-
-  public:
-    Toyota() {
-      // @todo
-    }
-
-}
-
-/**
- * \file Honda.cpp
- */
-
-class Honda: public Car {
-
-  public:
-    Honda() {
-      // @todo
-    }
-
-}
-
-/**
- * \file Ford.cpp
- */
-
-class Ford: public Car {
-
-  public:
-    Ford() {
-      // @todo
-    }
-
-}
-
-/**
- * \file Hyundai.cpp
- */
-
-class Hyundai: public Car {
-
-  public:
-    Hyundai() {
-      // @todo
-    }
-
-}
-
